move ModifyGroupState to otter.models.intents, and lint service.py
